revert: drop the address-form name adoption, keep the number mirror (TWO-40)
PR two-inc/prestashop-plugin#153 carried two separable changes. This commit
splits them and keeps only what held up.

UNAFFECTED - the precondition fix. `ajaxProcessSoleTraderTokens()` no longer
requires `id_address_invoice`; `resolveSoleTraderCountryIso()` resolves invoice
address -> posted ISO country -> delivery address; `TwoSoleTrader.fetchTokens()`
posts that country. Controller, `SoleTraderTokenPreconditionSpec.php`, its
registrations and the translations are untouched here, and its JS coverage is
carried over intact into `tests/js/sole-trader-mint-country.test.js`.

WITHDRAWN - the wide address-form adoption. The visible company-name write, the
`settledName` tag, the publish/persist backstop, `setCompanyIdHint()`, the
`sync*Visibility()` calls, the `companyName` event field, and the
`companySearchInAddressArea` config this PR added to `TwoCompanySearch` plus
both mount sites in `TwoCheckoutManager`. Two review rounds produced six
defects between them and were not converging: a country clobber through the
cookie writer's DOM-guessed country, an `id_address` pinned to the wrong
address, a mismatched name/number tag that made `hasConfirmedSelection()` lie,
and a nameless sole trader's enrolment destroyed by their own next keystroke.
`TwoCheckoutManager.js` is byte-identical to `staging` again.

KEPT, reduced - the organisation-number mirror. `two:sole-trader-ready` carries
`{company, companyid}`, and `onSoleTraderReady()` now does exactly three things:
stands down on `_destroyed`/`_manualEntry`, refuses a synthetic internal
identifier (the identification field is saved onto the address and can be
printed), and calls `writeOrganizationToAddressIdentifiers()`. That writer is
already gated on the merchant's "Autofill company address" setting
(`PS_TWO_ADDRESS_LOOKUP`) and announces no `input`/`change`, so there is no
cascade into the stale-selection clear. The gate is that setting and NOT where
company search is mounted - the setting happens to be forced off in the payment
tile, but that is a coincidence and is not relied on.

Docs reconciled to the reduced change, and `.ai/decisions.md` records what was
withdrawn and why, including that a true street/postcode/city autofill needs an
API contract confirmation because the buyer-autofill endpoint carries no address
payload.
