+++ AI_CONTEXT.md
@@ -99,10 +99,13 @@ Reliable B2B invoice checkout via Two, with:
     posted country -> delivery address and refuses if none resolves, with the
     registry check still the only authorisation gate (TWO-40 follow-up). On
     completion the module dispatches `two:sole-trader-ready` carrying the
-    enrolled name and organisation number; `TwoCompanySearch.js` adopts that
-    into the address form, but ONLY when it is mounted there
-    (`companySearchInAddressArea`) - the payment-tile placement must never
-    write into the address form
+    enrolled organisation number; `TwoCompanySearch.js` mirrors it into the
+    address `dni` field through `writeOrganizationToAddressIdentifiers()`, so
+    the gate is the merchant's "Autofill company address" setting
+    (`PS_TWO_ADDRESS_LOOKUP`) and NOT where company search is mounted. A
+    synthetic internal identifier is refused, and nothing is written into the
+    visible company field - see `.ai/decisions.md` for the wider adoption that
+    was withdrawn
 - `views/js/modules/TwoOptionalFields.js`
   - Optional buyer reference fields in the payment tile: mirrors each visible
     input into its hidden twin inside the payment form (the tile is a sibling
