+++ CHANGELOG.md
@@ -13,12 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The invoice-address precondition is replaced by a trust-ordered country resolver: the cart's invoice address first (never overridden by anything the request carries), then a posted `country` accepted only in exactly the two-upper-case-letter ISO shape, then the cart's delivery address. Nothing resolvable is still a refusal - now worded as a country problem rather than an address one
   - **The authorisation gate is unchanged**: minting still requires the registry to say the flow is available for a resolved country, re-evaluated server-side on every call. Accepting a posted country grants no privilege - the mint takes no country at all (the tokens are country-independent), the registry check still runs on the server, and the availability endpoint already answers any client-supplied country by design. So the posted value can only move the answer from "unresolved" to "the registry's own answer for a real country"
   - The country the browser posts is the same resolver the "Sole Trader" chip's own visibility is decided from, so the mint gate and the chip cannot disagree about which country the buyer is in
-- **A completed sole-trader enrolment left the address form's company field empty** (TWO-40 follow-up)
-  - The enrolled identity was persisted server-side and published in-memory, but never written into the form - so the buyer came back from the hosted signup to a blank company field, and the address they saved carried no company. `two:sole-trader-ready` now carries the enrolled name and organisation number, and the company-search control adopts them into the address form's `company` field (marked as autofilled, so a later selection can still tell it from typed input) and its identifier field through the existing address-lookup-gated writer
-  - **Address-form placement only.** The payment-tile placement must never write into the address form, so the placement is now passed explicitly to both mounts and is checked before any write - belt and braces alongside the tile mount's existing hardcoded `addressLookupEnabled: false`. The handler also stands down while the buyer is in manual entry, like every other write path in the control: enrolment is asynchronous and must not overwrite details being typed
-  - **The adoption no longer voids the enrolment it just saved** (adversarial review finding). Writing the company name announces `input`, whose handler clears the whole selection - including a `clearCompany` request and a blanked in-memory selection - unless the organisation number and its name tag already agree with the name in the field. For a buyer who had confirmed a registered company earlier in the session, that meant adopting the enrolment immediately undid it. The number and its tag are now written *before* the visible name, so there is nothing stale left to clear; the confirmed selection and the session cookie are re-asserted at the end of the handler as a backstop, not as the mechanism
-  - **The visible company field never receives the synthetic internal identifier** (adversarial review finding). The event's company *label* deliberately falls back to the organisation number, and for a sole trader with no trading name that number is the internal `TWO:`-prefixed identifier - which would then have been saved onto the address and printed on the invoice. The event now carries the real trading name separately, and only that value is written into the field the buyer sees; it is left untouched when there is no name. The identifier still travels wherever it legitimately has to, including the identifier field the credit decision reads
-  - Deliberately **no** street/postcode/city write-back: the buyer-autofill endpoint is consumed for the email, company name and organisation number only, on this plugin and the other platforms alike, so there is no verified address payload to write
+- **A completed sole-trader enrolment left the address form's identification field empty** (TWO-40 follow-up)
+  - The enrolled identity was persisted server-side and published in-memory, but nothing wrote the organisation number into the address form. `two:sole-trader-ready` now carries it, and the company-search control mirrors it into the address `dni` field through the existing writer - so it is marked as autofilled, `vat_number` is still never written, and no `input`/`change` is announced, which is what keeps it clear of the stale-selection clear
+  - **Gated on the merchant's "Autofill company address" setting** (`PS_TWO_ADDRESS_LOOKUP`), the same control every other address write a company capture performs already obeys, reached through the same gated writer rather than any check of its own. Deliberately *not* gated on where company search is mounted: that setting happens to be forced off when search moves to the payment tile, but the two are separate questions and the coincidence is not relied on
+  - A sole trader with no registered number enrols under a synthetic internal identifier, and that value is refused here: PrestaShop saves the identification field onto the address and can print it. The handler also stands down while the buyer is in manual entry, like every other write path in the control - enrolment is asynchronous and must not overwrite details being typed
+  - Deliberately **no** write into the visible company field, and **no** street/postcode/city write-back. The wider adoption was attempted and withdrawn (see `.ai/decisions.md`); the buyer-autofill endpoint is consumed for the email, company name and organisation number only, on this plugin and the other platforms alike, so there is no verified address payload to write
 - **The "Sole Trader" row in company search never appeared on the address-editor page, for any country** (TWO-40 follow-up, live bug on staging)
   - Root cause: `TwoSoleTrader.js`'s availability resolution (`refreshAvailability()`) early-returned whenever `.two-sole-trader` was absent from the page, before it ever resolved the billing country or fired the availability request. That container is rendered only by `paymentinfo.tpl` on the payment step - never on the address editor - so availability could never resolve anywhere else, however eligible the country. The container now only gates the enrolment prompt/status/error messaging it actually hosts; resolving and caching the availability answer no longer depends on it existing at all
 - **The availability round trip re-fired on every fresh page load, delaying the chip's first appearance** (Doug's own follow-up request)
